Struct magic::Cookie

pub struct Cookie { /* private fields */ }

Configuration of which CookieFlags and magic databases to use

Implementations

impl Cookie

pub fn file<P: AsRef<Path>>(&self, filename: P) -> Result<String, MagicError>

Returns a textual description of the contents of the filename

pub fn buffer(&self, buffer: &[u8]) -> Result<String, MagicError>

Returns a textual description of the contents of the buffer

pub fn set_flags(&self, flags: CookieFlags) -> Result<(), MagicError>

Sets the flags to use

Overwrites any previously set flags, e.g. those from load().

pub fn check<P: AsRef<Path>>(&self, filenames: &[P]) -> Result<(), MagicError>

Check the validity of entries in the database filenames

pub fn compile<P: AsRef<Path>>(&self, filenames: &[P]) -> Result<(), MagicError>

Compiles the given database filenames for faster access

The compiled files created are named from the basename of each file argument with ‘.mgc’ appended to it.

pub fn list<P: AsRef<Path>>(&self, filenames: &[P]) -> Result<(), MagicError>

Dumps all magic entries in the given database filenames in a human readable format

pub fn load<P: AsRef<Path>>(&self, filenames: &[P]) -> Result<(), MagicError>

Loads the given database filenames for further queries

Adds “.mgc” to the database filenames as appropriate.

Calling Cookie::load or Cookie::load_buffers replaces the previously loaded database/s.

Examples

let cookie = magic::Cookie::open(Default::default())?;

// Load the default database
cookie.load::<&str>(&[])?;

// Load databases from files
cookie.load(&vec!["data/tests/db-images-png", "data/tests/db-python"])?;

pub fn load_buffers(&self, buffers: &[&[u8]]) -> Result<(), MagicError>

Loads the given compiled databases for further queries

Databases need to be compiled with a compatible libmagic version.

This function can be used in environments where libmagic does not have direct access to the filesystem, but can access the magic database via shared memory or other IPC means.

Calling Cookie::load_buffers or Cookie::load replaces the previously loaded database/s.

pub fn open(flags: CookieFlags) -> Result<Cookie, MagicError>

Creates a new configuration, flags specify how other functions should behave

This does not load() any databases yet.